requirements.txt
Flask==2.3.0
PyJWT==2.8.0
bcrypt==4.0.1
Flask-Cors==4.0.0